active-projects/MM -- Area + Sub-Area Taxonomy/requirements.txt
pyyaml
notion-client
